validation-results/medium_rest_api_a127ee79_20251105_134521/code/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
bcrypt==4.0.1
PyJWT==2.8.0
pytest==7.4.0